requirements.txt
beautifulsoup4==4.12.0
numpy==1.20.1
opencv_python_headless==4.7.0.72
Pillow==9.5.0
plyer==2.1.0
pyzbar==0.1.9
qrcode==7.4.2
requests==2.25.1
streamlit==1.4.0
